Renamed Updatable to Monobehaviour
moved Destroyable logic to MonoBehaviour (tentative name), LifeCycleManager fully removes gameObject from its gameObjects hashmap/dictionary
